+++ src/Log/LogServiceProvider.php
@@ -2,11 +2,13 @@
 
 namespace Give\Log;
 
+use Give\Helpers\Hooks;
 use Give\ServiceProviders\ServiceProvider;
 use Give\Framework\Migrations\MigrationsRegister;
 use Give\Log\Migrations\CreateNewLogTable;
 use Give\Log\Migrations\MigrateExistingLogs;
 use Give\Log\Migrations\DeleteOldLogTables;
+use Give\Log\Helpers\Environment;
 
 /**
  * Class LogServiceProvider
@@ -30,6 +32,17 @@ class LogServiceProvider implements ServiceProvider {
 	 * @inheritdoc
 	 */
 	public function boot() {
+		$this->registerMigrations();
+		// Hook up
+		if ( Environment::isLogsPage() ) {
+			Hooks::addAction( 'admin_enqueue_scripts', Assets::class, 'enqueueScripts' );
+		}
+	}
+
+	/**
+	 * Register migration
+	 */
+	private function registerMigrations() {
 		give( MigrationsRegister::class )->addMigrations(
 			[
 				CreateNewLogTable::class,