+++ src/ServiceProviders/RestAPI.php
@@ -13,6 +13,8 @@ use Give\API\Endpoints\Reports\TopDonors;
 use Give\API\Endpoints\Reports\TotalDonors;
 use Give\API\Endpoints\Reports\TotalIncome;
 use Give\API\Endpoints\Reports\TotalRefunds;
+use Give\API\Endpoints\Logs\GetLogs;
+use Give\API\Endpoints\Logs\FlushLogs;
 use Give\API\RestRoute;
 
 class RestAPI implements ServiceProvider {
@@ -31,6 +33,8 @@ class RestAPI implements ServiceProvider {
 		TotalDonors::class,
 		TotalIncome::class,
 		TotalRefunds::class,
+		GetLogs::class,
+		FlushLogs::class,
 	];
 
 	/**